Handle out-of-order arrival of UTXOs, log stats in ledger_state
Because the message bus is asynchronous with multiple workers,
messages can be re-ordered.  This doesn't matter for stateless
things like the unpackers, but it does matter for tracking UTXOs,
obviously.

Since we assume this is all validated, we know that if we see
a spend which we don't yet know about, it must have been created
in a block which just hasn't arrived yet.  We therefore record it
in 'future_spends' and when it does arrive we just ignore it.

Currently outputting stats to Error pending rework of logging
levels.