Store UTXO lifetime, use for re-ordering
Waystage to handling rollbacks - will be less efficient than old
future_spend model since we have to look up in the entire set, not just
the future ones, but it's still a HashMap lookup and we will optimise later.

Note until we prune to 'k' the UTXO state will grow like Topsy because it will
record every UTXO for all time!