First attempt at handling rollback
When the rubber hits the road it raises all sorts of issues about handling rollbacks and re-ordering at the same time.  I'm beginning to think it may be impossible, and at the least it is moderately hard to reason about.  Having said
that it's vanishingly unlikely to happen in the current setup, since volatile blocks
are fetched one-by-one from upstream.

Given that UTXO updates are not particularly hard, maybe we should serialise
delta messages on input - basically hold any more than (max_number+1) in
abeyance until they become (max_number+1).