feat: initial utxo definition and code generation via cddl-codegen
This is very incomplete, the UTxO CDDL definition is not complete and
there are some issues we've encountered with cddl-codegen for generating
types around generics, for example. We've discussed a few potential
solutions to these problems in Slack.

Though it's not complete, we're getting everything merged in a state
where at least it's clear what needs to be worked on when we get back to
it, due to priority shifts at IO.

What is left todo in this specific commit:
  - Complete the UTxO definition
  - Once the CDDL is complete that defines the output from the Haskell
    node, write a simple tool that converts from that representation to
    the canonical representation.
  - Provide some initial test data to use (just testing SPO module to
    start)