Fix deposit timing, stakeless SPOs, epoch number logging
Deposits now repaid after snapshot taken, to avoid them changing
the stake

This revealed another problem, which is if an SPO withdraws all
its stake after a snapshot but still produces blocks, it would
have previously been ignored.  The early repayment of the deposit
previous masked this.

Switched logging of epochs in rewards and snapshots to be the
epoch just left - that is, the one that blocks are being counted for