Record chain of registration/deregistrations through snapshots
We need to know all effective (and not re-registered) deregistrations
that happened between 'go' and 'mark', and in early Shelley, the
ones in the current epoch until the rewards calc starts too.

Registration changes are now managed through the snapshot chain and
this makes it much easier to think about.