Add back GovernanceState with optional DRep,SPO subs
We need GovernanceState to update the protocol parameters even in
Shelley, but we don't want to include DReps.  This makes the DRep
and SPO distribution subscriptions optional which allows it to work
without them.

Also removed the blocks subscription hack from ParameterState since
we can now depend on Governance State being available - they come as a
couple.

Updated docs and configs to suit